set up the calibration
the calibration now works correctly and can help to make the mouse movement more specific to the person.

currently works well for horizontal movement but not vertical

vertical movement is implemented although quite noisy and innaccurate.
this is mostly due to when looking down the eye is covered by the eye lid so some other movement is going to need to be used for the vertical movement of the mouse